docs: document DB schema normalization and monster module completion
Update CHANGELOG.md and MONSTER_SYSTEM_RECONSTRUCTION.md to record the latest architectural updates and feature implementations:

- Database Normalization: Document the reorganization of SQL migrations and seeds into dedicated module directories (Accounts, Players, Equipment, WorldMap, Monsters) to improve maintainability.
- Monster Module: Track the addition of 4 new monster-specific database tables, the seeding of 78 confirmed asset entries, and the integration of `FramePaths` into asset contracts.
- Client Types: Record the creation of TypeScript types for monster rosters, battle states, and visual families.
- Backlog Update: Mark database schema, asset seeds, and catalog tracking tasks as complete in the monster system reconstruction documentation.

+++ server/Twelve.Application/Handlers/AuthHandler.cs
@@ -96,7 +96,7 @@ namespace Twelve.Application.Handlers
                 Phone        = phone,
                 Gender       = genderB,
             };
-            int accountId = await _accountRepository.CreateAsync(account);
+            long accountId = await _accountRepository.CreateAsync(account);
             _logger.LogInformation("[Register] Account created, Id={Id}", accountId);
 
             // ── Phản hồi thành công ───────────────────────────────────────────
